fixed

See merge request gp_groovy_git/travel2gather/back!14

+++ src/users/users.service.ts
@@ -253,12 +253,20 @@ export class UsersService {
     const updates = relations.flatMap(({ key, table, field }) => {
       const items = body[key as keyof UpdateProfileDto] as string[] | undefined;
       if (!items) return [];
-      return [
-        this.prismaService[table].deleteMany({ where: { userId } }),
-        this.prismaService[table].createMany({
-          data: items.map((item) => ({ userId, [field]: item })),
-        }),
-      ];
+      if (this.prismaService[table].count({ where: { userId } }) > 0) {
+        return [
+          this.prismaService[table].deleteMany({ where: { userId } }),
+          this.prismaService[table].createMany({
+            data: items.map((item) => ({ userId, [field]: item })),
+          }),
+        ];
+      } else {
+        return [
+          this.prismaService[table].createMany({
+            data: items.map((item) => ({ userId, [field]: item })),
+          }),
+        ];
+      }
     });
     await this.prismaService.$transaction(updates);
     return this.prismaService.profile.update({
